+++ src/view/src/rocprofvis_timeline_view.h
@@ -127,7 +127,7 @@ private:
     std::pair<double, double>           m_highlighted_region;
     SettingsManager&                    m_settings;
     double                              m_last_data_req_view_time_offset_ns;
-    int                                 m_artificial_scrollbar_height;
+    float                               m_artificial_scrollbar_height;
     ImVec2                              m_graph_size;
     double                              m_grid_interval_ns;
     int                                 m_grid_interval_count;
